moved effect controller to separate file
moved effect controller to separate file
